Split LedController from tests
