+++ td/challenge260/Instruction.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <vector>
+
 enum class InstructionType
 {
   LdA,
