+++ td/challenge260/LedControllerTest.cpp
@@ -1,66 +1,6 @@
 #include "gmock/gmock.h"
 
-#include <sstream>
-#include <bitset>
-
-#include "Instruction.hpp"
-
-class LedController
-{
-public:
-  explicit LedController(std::ostream &stream) :
-    out{stream}
-  {
-  }
-
-  void runProgram(const Instructions &instructions)
-  {
-    for (const auto &instruction : instructions)
-    {
-      runInstruction(instruction);
-    }
-  }
-
-private:
-  std::ostream &out;
-  std::string ledState{getLedStateFromInteger(0)};
-
-  void runInstruction(const Instruction & instruction)
-  {
-    switch (instruction.type)
-    {
-      case InstructionType::OutA:
-        out << ledState;
-        break;
-      case InstructionType::LdA:
-        ledState = getLedStateFromInteger(instruction.value);
-        break;
-    }
-  }
-
-  static std::string getLedStateFromInteger(unsigned value)
-  {
-    const std::bitset<8> bitValue{value};
-    std::stringstream stream;
-    stream << bitValue;
-    auto text = stream.str();
-    std::transform(text.cbegin(), text.cend(), text.begin(), [](const auto character)
-    {
-      if (character == '1')
-      {
-        constexpr auto ledOnChar = '*';
-        return ledOnChar;
-      }
-      else
-      {
-        constexpr auto ledOffChar = '.';
-        return ledOffChar;
-      }
-    });
-    return text + '\n';
-  }
-
-};
+#include "LedController.hpp"
 
 static constexpr Instruction createInstructionWithZeroValue(InstructionType type)
 {
